Add options to configure behavior in case of OOM.
OOM situations should not be dropped on the "always fail" floor. Instead, give a chance that
a request may be served anyway, even if it will be delayed by a OOM situation.

The httpd will sleep the given amount of time in a single attempts then will try again. If it
will hit the same status, it will sleep again, but with attempts counter increment. If OOM
situation will go away, then this counter is reset and anything is back to normal again.
But if counter will hit the "maximum attempts" limit, the fail will be loud and complete.

By default this behavior is disabled, and failure is immediate, as it was before.

New options:
-O oom_timer: sets a sleep timer, which spends time in hope that host's OOM situation will go away.
-O oom_max_attempts: sets number of maximum attempts after which httpd will completely fail.

Both options must be set to nonzero values to enable the behavior.