Implement directory listing sorting.
This was in NOTTODO, however I considered this a mandatory feature.
I tired using httpd without sorted long directory lists.
The memory taken by the list is not so much, and even if OOM will be
triggered, then I implemented a "idxfrom=" and "idxto=" parameters to
list very big directories partially without loading the full
index data into memory.

The algorithm is simple as everywhere:
- Actually get unsorted list from OS,
- Sort it with qsort,
- Output sorted list, format it.

There are parameters to control sorting, not currently exposed as a html UI elements:
- "sortby=" can take "name", "type", "size", "owner"/"uid", "group"/"gid", "time" or "none".
The names are self descriptory.
- "rsort=1" simply reverses sorting order.

By default it sorts the list by file name. Even if other type of sorting is applied,
unsorted elements which cannot be sorted with applied type are then sorted by name.